bump keycloak subchart version
We were seeing https://github.com/bitnami/charts/issues/18933 because of
inconsistent version of bitnami/common in the dependency tree:

naavre
├─ vrepaas 0.5.7
│  └─ bitnami/postgresql 12.12.10
│     └─ bitnami/common 2.12.01
└─ keycloak 15.1.6
   ├─ bitnami/common 2.4.0
   └─ bitnami/postgresql 12.5.8
      └─ bitnami/common 2.4.0

Bumping keycloak to 16.x.x gives consistent version of bitnami/common:

naavre
├─ vrepaas 0.5.7
│  └─ bitnami/postgresql 12.12.10
│     └─ bitnami/common 2.12.01
└─ keycloak 16.1.7
   ├─ bitnami/common 2.12.0
   └─ bitnami/postgresql 12.12.9
      └─ bitnami/common 2.12.0

Although this fixes the issue now, this doesn’t seem very sustainable.
In the future, we will need to carefuly keep versions of the vrepaas and
keycloak charts aligned, making sure that they use the same
bitnami/common dependency. If anything, this is a case for not using
bitnami charts when it can be avoided.